pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "digital-caffeine"
version = "1.0.0"
description = "Keep your Windows machine awake - CLI and system tray utility"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Blake Bratcher" },
]
keywords = ["windows", "keep-awake", "caffeine", "sleep", "utility"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pystray>=0.19",
    "Pillow>=10.0",
    "textual>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
    "ruff>=0.4",
]

[project.scripts]
caffeine = "digital_caffeine.cli:cli"

[project.urls]
Homepage = "https://github.com/blakebratcher/digital-caffeine"
Repository = "https://github.com/blakebratcher/digital-caffeine"
Issues = "https://github.com/blakebratcher/digital-caffeine/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/digital_caffeine"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
